Move interpreter test methods to test module
