+++ evm_arithmetization/src/cpu/kernel/interpreter.rs
@@ -1,7 +1,6 @@
 //! An EVM interpreter for testing and debugging purposes.
 
 use core::cmp::Ordering;
-use core::ops::Range;
 use std::collections::{BTreeSet, HashMap};
 
 use anyhow::anyhow;
@@ -12,9 +11,7 @@ use plonky2::field::types::Field;
 use crate::byte_packing::byte_packing_stark::BytePackingOp;
 use crate::cpu::columns::CpuColumnsView;
 use crate::cpu::kernel::aggregator::KERNEL;
-use crate::cpu::kernel::constants::context_metadata::ContextMetadata;
 use crate::cpu::kernel::constants::global_metadata::GlobalMetadata;
-use crate::cpu::kernel::constants::txn_fields::NormalizedTxnField;
 use crate::generation::debug_inputs;
 use crate::generation::mpt::load_all_mpts;
 use crate::generation::rlp::all_rlp_prover_inputs_reversed;
@@ -27,9 +24,7 @@ use crate::keccak_sponge::keccak_sponge_stark::KeccakSpongeOp;
 use crate::memory::segments::Segment;
 use crate::util::h2u;
 use crate::witness::errors::ProgramError;
-use crate::witness::memory::{
-    MemoryAddress, MemoryContextState, MemoryOp, MemoryOpKind, MemorySegmentState,
-};
+use crate::witness::memory::{MemoryAddress, MemoryOp, MemoryOpKind, MemorySegmentState};
 use crate::witness::operation::Operation;
 use crate::witness::state::RegistersState;
 use crate::witness::transition::{
@@ -368,61 +363,6 @@ impl<F: Field> Interpreter<F> {
             .segments[Segment::Code.unscale()]
     }
 
-    fn code_slice(&self, n: usize) -> Vec<u8> {
-        let pc = self.generation_state.registers.program_counter;
-        self.code().content[pc + 1..pc + n + 1]
-            .iter()
-            .map(|u256| u256.unwrap_or_default().byte(0))
-            .collect::<Vec<_>>()
-    }
-
-    pub(crate) fn get_txn_field(&self, field: NormalizedTxnField) -> U256 {
-        // These fields are already scaled by their respective segment.
-        self.generation_state.memory.contexts[0].segments[Segment::TxnFields.unscale()]
-            .get(field.unscale())
-    }
-
-    pub(crate) fn set_txn_field(&mut self, field: NormalizedTxnField, value: U256) {
-        // These fields are already scaled by their respective segment.
-        self.generation_state.memory.contexts[0].segments[Segment::TxnFields.unscale()]
-            .set(field.unscale(), value);
-    }
-
-    pub(crate) fn get_txn_data(&self) -> Vec<U256> {
-        self.generation_state.memory.contexts[0].segments[Segment::TxnData.unscale()].content()
-    }
-
-    pub(crate) fn get_context_metadata_field(&self, ctx: usize, field: ContextMetadata) -> U256 {
-        // These fields are already scaled by their respective segment.
-        self.generation_state.memory.contexts[ctx].segments[Segment::ContextMetadata.unscale()]
-            .get(field.unscale())
-    }
-
-    pub(crate) fn set_context_metadata_field(
-        &mut self,
-        ctx: usize,
-        field: ContextMetadata,
-        value: U256,
-    ) {
-        // These fields are already scaled by their respective segment.
-        self.generation_state.memory.contexts[ctx].segments[Segment::ContextMetadata.unscale()]
-            .set(field.unscale(), value)
-    }
-
-    pub(crate) fn get_global_metadata_field(&self, field: GlobalMetadata) -> U256 {
-        // These fields are already scaled by their respective segment.
-        let field = field.unscale();
-        self.generation_state.memory.contexts[0].segments[Segment::GlobalMetadata.unscale()]
-            .get(field)
-    }
-
-    pub(crate) fn set_global_metadata_field(&mut self, field: GlobalMetadata, value: U256) {
-        // These fields are already scaled by their respective segment.
-        let field = field.unscale();
-        self.generation_state.memory.contexts[0].segments[Segment::GlobalMetadata.unscale()]
-            .set(field, value)
-    }
-
     pub(crate) fn set_global_metadata_multi_fields(&mut self, metadata: &[(GlobalMetadata, U256)]) {
         for &(field, value) in metadata {
             let field = field.unscale();
@@ -431,141 +371,10 @@ impl<F: Field> Interpreter<F> {
         }
     }
 
-    pub(crate) fn get_trie_data(&self) -> Vec<U256> {
-        self.generation_state.memory.contexts[0].segments[Segment::TrieData.unscale()]
-            .content
-            .iter()
-            .filter_map(|&elt| elt)
-            .collect::<Vec<_>>()
-    }
-
-    pub(crate) fn get_trie_data_mut(&mut self) -> &mut Vec<Option<U256>> {
-        &mut self.generation_state.memory.contexts[0].segments[Segment::TrieData.unscale()].content
-    }
-
-    pub(crate) fn get_memory_segment(&self, segment: Segment) -> Vec<U256> {
-        if self.is_preinitialized_segment(segment.unscale()) {
-            let total_len = self.generation_state.memory.contexts[0].segments[segment.unscale()]
-                .content
-                .len();
-            let get_vals = |opt_vals: &[Option<U256>]| {
-                opt_vals
-                    .iter()
-                    .map(|&elt| match elt {
-                        Some(val) => val,
-                        None => U256::zero(),
-                    })
-                    .collect::<Vec<U256>>()
-            };
-            let mut res = get_vals(
-                &self
-                    .generation_state
-                    .memory
-                    .get_preinitialized_segment(segment)
-                    .expect("The segment should be in the preinitialized segments.")
-                    .content,
-            );
-            let init_len = res.len();
-            res.extend(&get_vals(
-                &self.generation_state.memory.contexts[0].segments[segment.unscale()].content
-                    [init_len..],
-            ));
-            res
-        } else {
-            self.generation_state.memory.contexts[0].segments[segment.unscale()].content()
-        }
-    }
-
-    pub(crate) fn get_memory_segment_bytes(&self, segment: Segment) -> Vec<u8> {
-        let content = self.get_memory_segment(segment);
-        content.iter().map(|x| x.low_u32() as u8).collect()
-    }
-
-    pub(crate) fn get_current_general_memory(&self) -> Vec<U256> {
-        self.generation_state.memory.contexts[self.context()].segments
-            [Segment::KernelGeneral.unscale()]
-        .content()
-    }
-
-    pub(crate) fn get_kernel_general_memory(&self) -> Vec<U256> {
-        self.get_memory_segment(Segment::KernelGeneral)
-    }
-
-    pub(crate) fn get_rlp_memory(&self) -> Vec<u8> {
-        self.get_memory_segment_bytes(Segment::RlpRaw)
-    }
-
-    pub(crate) fn set_current_general_memory(&mut self, memory: Vec<U256>) {
-        let context = self.context();
-        self.generation_state.memory.contexts[context].segments[Segment::KernelGeneral.unscale()]
-            .content = memory.iter().map(|&val| Some(val)).collect();
-    }
-
-    pub(crate) fn set_memory_segment(&mut self, segment: Segment, memory: Vec<U256>) {
-        self.generation_state.memory.contexts[0].segments[segment.unscale()].content =
-            memory.iter().map(|&val| Some(val)).collect();
-    }
-
-    pub(crate) fn set_memory_segment_bytes(&mut self, segment: Segment, memory: Vec<u8>) {
-        self.generation_state.memory.contexts[0].segments[segment.unscale()].content = memory
-            .into_iter()
-            .map(|val| Some(U256::from(val)))
-            .collect();
-    }
-
-    pub(crate) fn extend_memory_segment(&mut self, segment: Segment, memory: &[Option<U256>]) {
-        self.generation_state.memory.contexts[0].segments[segment.unscale()]
-            .content
-            .extend(memory);
-    }
-
-    pub(crate) fn extend_memory_segment_bytes(&mut self, segment: Segment, memory: Vec<u8>) {
-        self.generation_state.memory.contexts[0].segments[segment.unscale()]
-            .content
-            .extend(
-                memory
-                    .into_iter()
-                    .map(|elt| Some(U256::from(elt)))
-                    .collect::<Vec<_>>(),
-            );
-    }
-
-    pub(crate) fn set_rlp_memory(&mut self, rlp: Vec<u8>) {
-        self.set_memory_segment_bytes(Segment::RlpRaw, rlp)
-    }
-
-    pub(crate) fn set_code(&mut self, context: usize, code: Vec<u8>) {
-        assert_ne!(context, 0, "Can't modify kernel code.");
-        while self.generation_state.memory.contexts.len() <= context {
-            self.generation_state
-                .memory
-                .contexts
-                .push(MemoryContextState::default());
-        }
-        self.generation_state.memory.set(
-            MemoryAddress::new(
-                context,
-                Segment::ContextMetadata,
-                ContextMetadata::CodeSize.unscale(),
-            ),
-            code.len().into(),
-        );
-        self.generation_state.memory.contexts[context].segments[Segment::Code.unscale()].content =
-            code.into_iter().map(|val| Some(U256::from(val))).collect();
-    }
-
     pub(crate) fn set_memory_multi_addresses(&mut self, addrs: &[(MemoryAddress, U256)]) {
         for &(addr, val) in addrs {
             self.generation_state.memory.set(addr, val);
         }
-    }
-
-    pub(crate) fn set_jumpdest_analysis_inputs(&mut self, jumps: HashMap<usize, BTreeSet<usize>>) {
-        self.generation_state.set_jumpdest_analysis_inputs(jumps);
-    }
-
-    pub(crate) fn incr(&mut self, n: usize) {
-        self.generation_state.registers.program_counter += n;
     }
 
     pub(crate) fn stack(&self) -> Vec<U256> {
@@ -601,92 +410,6 @@ impl<F: Field> Interpreter<F> {
             .content
     }
 
-    pub(crate) fn extract_kernel_memory(self, segment: Segment, range: Range<usize>) -> Vec<U256> {
-        let mut output: Vec<U256> = Vec::with_capacity(range.end);
-        for i in range {
-            let term = self
-                .generation_state
-                .memory
-                .get_with_init(MemoryAddress::new(0, segment, i));
-            output.push(term);
-        }
-        output
-    }
-
-    // Actually pushes in memory. Only used for tests.
-    #[cfg(test)]
-    pub(crate) fn push(&mut self, x: U256) -> Result<(), ProgramError> {
-        use crate::cpu::stack::MAX_USER_STACK_SIZE;
-
-        if !self.is_kernel() && self.stack_len() >= MAX_USER_STACK_SIZE {
-            return Err(ProgramError::StackOverflow);
-        }
-        if self.stack_len() > 0 {
-            let top = self
-                .stack_top()
-                .expect("The stack is checked to be nonempty");
-            let cur_len = self.stack_len();
-            let stack_addr = MemoryAddress::new(self.context(), Segment::Stack, cur_len - 1);
-            self.generation_state.memory.set(stack_addr, top);
-        }
-        self.generation_state.registers.stack_top = x;
-        self.generation_state.registers.stack_len += 1;
-
-        Ok(())
-    }
-
-    /// Actually popping the memory. Only used in tests.
-    #[cfg(test)]
-    pub(crate) fn pop(&mut self) -> Result<U256, ProgramError> {
-        use crate::witness::util::stack_peek;
-
-        let result = stack_peek(&self.generation_state, 0);
-
-        if self.stack_len() > 1 {
-            let top = stack_peek(&self.generation_state, 1)?;
-            self.generation_state.registers.stack_top = top;
-        }
-        self.generation_state.registers.stack_len -= 1;
-
-        result
-    }
-
-    fn offset_name(&self) -> String {
-        KERNEL.offset_name(self.generation_state.registers.program_counter)
-    }
-
-    fn offset_label(&self) -> Option<String> {
-        KERNEL.offset_label(self.generation_state.registers.program_counter)
-    }
-
-    pub(crate) fn get_jumpdest_bit(&self, offset: usize) -> U256 {
-        if self.generation_state.memory.contexts[self.context()].segments
-            [Segment::JumpdestBits.unscale()]
-        .content
-        .len()
-            > offset
-        {
-            // Even though we are in the interpreter, `JumpdestBits` is not part of the
-            // preinitialized segments, so we don't need to carry out the additional checks
-            // when get the value from memory.
-            self.generation_state.memory.get_with_init(MemoryAddress {
-                context: self.context(),
-                segment: Segment::JumpdestBits.unscale(),
-                virt: offset,
-            })
-        } else {
-            0.into()
-        }
-    }
-
-    pub(crate) fn get_jumpdest_bits(&self, context: usize) -> Vec<bool> {
-        self.generation_state.memory.contexts[context].segments[Segment::JumpdestBits.unscale()]
-            .content
-            .iter()
-            .map(|x| x.unwrap_or_default().bit(0))
-            .collect()
-    }
-
     pub(crate) fn add_jumpdest_offset(&mut self, offset: usize) {
         if let Some(jumpdest_table) = self
             .jumpdest_table
@@ -717,19 +440,8 @@ impl<F: Field> Interpreter<F> {
         self.generation_state.registers.is_kernel
     }
 
-    pub(crate) fn set_is_kernel(&mut self, is_kernel: bool) {
-        self.generation_state.registers.is_kernel = is_kernel
-    }
-
     pub(crate) const fn context(&self) -> usize {
         self.generation_state.registers.context
-    }
-
-    pub(crate) fn set_context(&mut self, context: usize) {
-        if context == 0 {
-            assert!(self.is_kernel());
-        }
-        self.generation_state.registers.context = context;
     }
 
     /// Writes the encoding of 0 at @ENCODED_EMPTY_NODE_ADDR.
@@ -747,7 +459,7 @@ impl<F: Field> Interpreter<F> {
             .insert_preinitialized_segment(segment, values);
     }
 
-    fn is_preinitialized_segment(&self, segment: usize) -> bool {
+    pub(crate) fn is_preinitialized_segment(&self, segment: usize) -> bool {
         self.generation_state
             .memory
             .is_preinitialized_segment(segment)
@@ -868,7 +580,6 @@ impl<F: Field> State<F> for Interpreter<F> {
         // will overwrite these ones during the op generation.
         if let Some(special_len) = get_op_special_length(op) {
             let special_len_f = F::from_canonical_usize(special_len);
-            let diff = row.stack_len - special_len_f;
             if (generation_state.stack().len() != special_len) {
                 // If the `State` is an interpreter, we cannot rely on the row to carry out the
                 // check.
@@ -1121,7 +832,6 @@ fn get_mnemonic(opcode: u8) -> &'static str {
 
 #[cfg(test)]
 mod tests {
-
     use ethereum_types::U256;
     use plonky2::field::goldilocks_field::GoldilocksField as F;
